requirements.txt
asgiref
CacheControl
cachetools
certifi
cffi
chardet
cryptography
defusedxml
Deprecated
Django~=3.2.9
django-filter~=21.1
django-phonenumber-field
django-solo
djangorestframework~=3.12.4
drf-firebase-token-auth
firebase-admin
gcloud
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
gunicorn
httplib2
idna
install
jwcrypto
Markdown
msgpack
oauth2client
oauthlib
packaging
Pillow
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pycryptodome
PyJWT~=2.3.0
pyparsing
Pyrebase4
python-jwt
python3-openid
pytz
requests
requests-oauthlib
requests-toolbelt
rest-firebase-auth
rsa
six
social-auth-app-django
social-auth-core
sqlparse
uritemplate
urllib3
wrapt
djangorestframework-jwt~=1.11.0
pyotp~=2.6.0
twilio~=7.3.0